# internal/model/tag.go
package model

import "github.com/testercc/blog-service/pkg/app"

// 创建标签model

type Tag struct {
	*Model
	Name  string `json:"name"`
	State uint8  `json:"state"`
}

type TagSwagger struct {
	List  []*Tag
	Pager *app.Pager
}

func (t Tag) TableName() string {
	return "blog_tag"
}